pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flower-store"
version = "0.0.1"
authors = [
    {name = "David Rambo", email = "davidrambo@mailfence.com"},
]
description = "Store for selling flowers built on Flask."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography",
    "elasticsearch",
    "flask",
    "flask-admin",
    "flask-assets",
    "flask-login",
    "flask-migrate",
    "flask-sqlalchemy",
    "flask-wtf",
    "gunicorn",
    "pillow",
    "pymysql",
    "psycopg2-binary",
    "pytailwindcss",
    "python-dotenv",
]

[project.urls]
homepage = "http://github.com/DavidRambo/flower-store"
repository = "http://github.com/DavidRambo/flower-store"

[project.optional-dependencies]
dev = [
    "black",
    "coverage",
    "djlint",
    "pytest",
]

[tool.coverage.paths]
source = ["src"]

[tool.coverage.run]
branch = true
source = ["flower_store"]

[tool.coverage.report]
show_missing = true
fail_under = 100

[tool.djlint]
indent=2
profile="jinja"
# no_function_formatting=true

